Refactor Character Animations
