+++ systems/character/character.go
@@ -10,18 +10,6 @@ import (
 	"github.com/eth0net/magicgame/systems/speed"
 )
 
-// CharacterAnimations contains animations for Character.
-var CharacterAnimations = []*common.Animation{
-	{Name: AnimationMoveUp, Frames: []int{9, 10, 11}, Loop: true},
-	{Name: AnimationMoveDown, Frames: []int{0, 1, 2}, Loop: true},
-	{Name: AnimationMoveLeft, Frames: []int{3, 4, 5}, Loop: true},
-	{Name: AnimationMoveRight, Frames: []int{6, 7, 8}, Loop: true},
-	{Name: AnimationStopUp, Frames: []int{10}, Loop: true},
-	{Name: AnimationStopDown, Frames: []int{1}, Loop: true},
-	{Name: AnimationStopLeft, Frames: []int{4}, Loop: true},
-	{Name: AnimationStopRight, Frames: []int{7}, Loop: true},
-}
-
 // Character entity is a game character.
 type Character struct {
 	ecs.BasicEntity
