--- systems/character/constants.go
@@ -1,13 +1,0 @@
-package character
-
-// Animation names.
-const (
-	AnimationMoveUp    string = "moveUp"
-	AnimationMoveDown  string = "moveDown"
-	AnimationMoveLeft  string = "moveLeft"
-	AnimationMoveRight string = "moveRight"
-	AnimationStopUp    string = "stopUp"
-	AnimationStopDown  string = "stopDown"
-	AnimationStopLeft  string = "stopLeft"
-	AnimationStopRight string = "stopRight"
-)
